requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
google-genai>=1.0
python-dotenv>=1.0
